requirements.txt
anyio==3.6.2
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
attrs==22.1.0
backcall==0.2.0
bayesian-optimization==1.4.2
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-default-group==1.2.2
cloup==0.13.1
colorama==0.4.6
colour==0.1.5
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
glcontext==2.3.7
greenlet==2.0.1
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isosurfaces==0.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
manim==0.17.3
ManimPango==0.4.3
mapbox-earcut==1.0.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
moderngl==5.8.2
moderngl-window==2.4.4
multipledispatch==0.6.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
playwright==1.28.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
pycairo==1.23.0
pycparser==2.21
pydub==0.25.1
pyee==9.0.4
pygame==2.4.0
pyglet==2.0.7
Pygments==2.13.0
pyparsing==3.0.9
pyrr==0.10.3
pyrsistent==0.19.2
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.30.0
rich==13.3.5
scikit-learn==1.1.3
scipy==1.9.3
screeninfo==0.8.1
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
skia-pathops==0.7.4
sniffio==1.3.0
soupsieve==2.3.2.post1
srt==3.5.3
stack-data==0.6.2
svgelements==1.9.5
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.1
tqdm==4.65.0
traitlets==5.6.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==2.0.2
virtualenv==20.17.0
watchdog==2.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
windows-curses==2.3.1
